backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "openai>=1.106.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.35.0",
]
